requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-genai==1.50.1
python-dotenv==1.0.1
